move java thermal impls to another module
