+++ Android/.idea/gradle.xml
@@ -12,6 +12,7 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/../Packages/MobileSupportStorage/Runtime/Plugins/Android/MobileSupportStorage.androidlib" />
+            <option value="$PROJECT_DIR$/../Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib" />
           </set>
         </option>
       </GradleProjectSettings>
