+++ Android/app/build.gradle
@@ -29,6 +29,7 @@ android {
 
 dependencies {
     implementation project(':MobileSupportStorage.androidlib')
+    implementation project(':MobileSupportThermal.androidlib')
 
     implementation 'androidx.appcompat:appcompat:1.3.0'
     implementation 'com.google.android.material:material:1.4.0'
