+++ Packages/MobileSupportStorage/Runtime/Plugins/Android/MobileSupportStorage.androidlib/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="jp.co.cyberagent.unitysupport">
+    package="jp.co.cyberagent.unitysupport.storage">
 
 </manifest>
