+++ Packages/MobileSupportStorage/Runtime/Plugins/Android/MobileSupportStorage.androidlib/src/main/java/jp/co/cyberagent/unitysupport/storage.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60140edef47ec42b3aacbf17f3a948e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
