+++ Packages/MobileSupportStorage/Runtime/Plugins/Android/MobileSupportStorage.androidlib/src/main/java/jp/co/cyberagent/unitysupport/storage/Storage.java
@@ -1,4 +1,4 @@
-package jp.co.cyberagent.unitysupport;
+package jp.co.cyberagent.unitysupport.storage;
 
 import android.annotation.TargetApi;
 import android.content.Context;
