+++ Packages/MobileSupportStorage/Runtime/Plugins/Android/MobileSupportStorage.androidlib/src/main/java/jp/co/cyberagent/unitysupport/storage/Storage.java.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 15c5681ad01c344fdaf5a57587c10947
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
