+++ Packages/MobileSupportThermal/Runtime/Plugins.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb8823f17bd6746b18597b94d3293e68
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
