+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/.gitignore
@@ -1,0 +1,1 @@
+/build
