+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/build.gradle.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dfa90b6db2b8d4411b5d0f024f139c94
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
