+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/proguard-rules.pro.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f6085a2c6e52e413082684d3795f84c1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
