+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f1caefc374814281a9a74c5af9f93b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
