+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f29b74a2f7e134f4192d24f046aba665
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
