+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/AndroidManifest.xml.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3ea67daa9b5cb4499b5faab34b6b387c
+guid: d346eb4eacd3341a38e319e677825122
 DefaultImporter:
   externalObjects: {}
   userData:
