+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/java.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3310b9d8979004689b23dfbb8f78b907
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
