+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/java/jp.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 77769b23ab9654d80a3c81285dd55e18
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
