+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/java/jp/co.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d42871230b5442f9b3bb7a4119fa24d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
