+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/java/jp/co/cyberagent.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2c0909750eee344ae924590b13be3b3f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
