+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/java/jp/co/cyberagent/unitysupport.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 1fb1f4d6320b44cbf851ce8f33de4d3d
+guid: a546f0ae681f54c168be14faf4115d1a
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
