+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/java/jp/co/cyberagent/unitysupport/thermal/BatteryChangedBroadcastReceiver.java.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 609605332f45142198d035f82c94ff65
+guid: 73df30bc6124e4847aa38d8017fc800c
 DefaultImporter:
   externalObjects: {}
   userData:
