+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/java/jp/co/cyberagent/unitysupport/thermal/BatteryTemperatureReceiver.java
@@ -1,4 +1,4 @@
-package jp.co.cyberagent.unitysupport;
+package jp.co.cyberagent.unitysupport.thermal;
 
 public interface BatteryTemperatureReceiver {
     void onReceiveBatteryTemperature(int level);
