+++ Packages/MobileSupportThermal/Runtime/Plugins/Android/MobileSupportThermal.androidlib/src/main/java/jp/co/cyberagent/unitysupport/thermal/BatteryTemperatureReceiver.java.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3f377d35e5a7a43deb9adec515c50655
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
